Add files via upload
TP 5 terminé

+++ Annuaire_Etd_poo/controller/EmployeController.php
@@ -12,7 +12,8 @@ class EmployeController {
 
     public function index($notification = '') {
         $data['notification'] = $notification;
-    	$data['employes'] = $this->employes->getAllEmploye();
+        $data['employes'] = $this->employes->getAllEmploye();
+        $data['categorie'] = $this->employes->Categorie();
     	include 'view/employe/index.php';
         die;
     }
@@ -48,6 +49,7 @@ class EmployeController {
             
             if (empty($errors)) {
                 $add = $this->employes->add($_POST);
+                
                 if ($add) {
                     $msg = "L'employé ".$_POST['prenom'].$_POST['nom']." a été ajouté!";
                 } 
@@ -57,20 +59,41 @@ class EmployeController {
                 $this->index($msg); // Redirection vers l'index
             }
         }
+        $data['categorie'] = $this->employes->categorie();
         include 'view/employe/add.php';
     }
 
 
     public function delete()
     {
-        $del = $this->employes->delete($_GET['id']);
-        if ($del) {
-            $msg = "L'employé ". $_GET['id']." a été supprimé.";
-        } 
-        else {
-            $msg = "Impossible de supprimer l'employé!";
+        session_start();
+        $type = $_SESSION['type_compte'];
+        if (empty($_SESSION)){
+            echo "<script type='text/javascript'>alert('Vous devez être connecté pour accéder à cette fonction'); 
+            window.location.href='?ctrl=employe&mth=login';</script>";
+            //echo "<script type='text/javascript'> alert('test'); </script>";
+            //header("Location: ?ctrl=Accueil&mth=index");
+            //exit();
+            
         }
-        $this->index($msg); // Redirection vers l'index
+        
+        else if ($type == 0){
+            echo "<script type='text/javascript'>alert('Vous devez être administrateur pour accéder à cette fonction'); 
+            window.location.href='?ctrl=employe';</script>";
+        }
+
+        else {
+            $del = $this->employes->delete($_GET['id']);
+            if ($del) {
+                $msg = "L'employé ". $_GET['id']." a été supprimé.";
+            } 
+            else {
+                $msg = "Impossible de supprimer l'employé!";
+            }
+            $this->index($msg); // Redirection vers l'index
+        }
+
+        
     }
 
     public function edit()
@@ -100,6 +123,7 @@ class EmployeController {
 
         }
         $employe = $this->employes->getEmploye($_GET['id']);
+        $data['categorie'] = $this->employes->categorie();
         include 'view/employe/edit.php';
     
     }
@@ -142,6 +166,83 @@ class EmployeController {
             header("Location: ?ctrl=Accueil&mth=index");
             
             
+        }
+    }
+
+    public function trie()
+    {
+        if (isset($_POST['trie'])){
+            $data['employes'] = $this->employes->trie($_POST['categorie']);
+            include 'view/employe/trie.php';
+            
+        }
+    }
+
+    public function ordre()
+    {
+        if (!isset($_SESSION)){
+            session_start();
+        }
+        $data['employes'] = $this->employes->ordre($_SESSION['categorie'],$_GET['by'],$_GET['AD']);
+        include 'view/employe/trie.php';  
+    }
+
+    public function categorie()
+    {
+        $data['categorie'] = $this->employes->categorie();
+        include 'view/employe/categorie.php';
+    }
+
+    public function editCat()
+    {
+        $errors = array();
+        if (isset($_POST['submit'])) {
+
+            $editCat = $this->employes->editCat($_POST,$_GET['id']);
+            
+
+            if ($editCat) {
+                $msg = "La catégorie ". $_GET['id']." a été modifiée.";
+            } 
+            else {
+                $msg = "Impossible de modifier la catégorie!";
+            }
+            $this->index($msg); // Redirection vers l'index
+
+        }
+        $categorie = $this->employes->getCategorie($_GET['id']);
+        include 'view/employe/editCat.php';
+    
+    }
+
+    public function deleteCat()
+    {
+        session_start();
+        $type = $_SESSION['type_compte'];
+        if (empty($_SESSION)){
+            echo "<script type='text/javascript'>alert('Vous devez être connecté pour accéder à cette fonction'); 
+            window.location.href='?ctrl=employe&mth=login';</script>";
+            //echo "<script type='text/javascript'> alert('test'); </script>";
+            //header("Location: ?ctrl=Accueil&mth=index");
+            //exit();
+            
+        }
+        
+        else if ($type == 0){
+            echo "<script type='text/javascript'>alert('Vous devez être administrateur pour accéder à cette fonction'); 
+            window.location.href='?ctrl=employe';</script>";
+        }
+
+        else {
+
+            $del = $this->employes->deleteCat($_GET['id']);
+            if ($del) {
+                $msg = "La catégorie ". $_GET['id']." a été supprimé.";
+            } 
+            else {
+                $msg = "Impossible de supprimer la catégorie!";
+            }
+            $this->index($msg); // Redirection vers l'index
         }
     }
 }
